Fixed function and module addresses on Linux.
Functions are listed with "nm".  Functions that can receive bpftrace
uprobes are then tagged through the m_Probe field in Function.

+++ OrbitCore/OrbitFunction.h
@@ -141,7 +141,7 @@ public:
     std::vector<FunctionParam>     m_Params;
     std::vector<Argument>          m_ArgInfo;
     Pdb*                           m_Pdb;
-    unsigned long long             m_NameHash;
+    uint64_t                       m_NameHash;
     OrbitType                      m_OrbitType;
     std::shared_ptr<FunctionStats> m_Stats;
